modifico tipo de dato en metodos

+++ src/main/java/com/uade/e_commerce/dto/CarritoItemResponseDTO.java
@@ -16,7 +16,7 @@ public class CarritoItemResponseDTO {
     private Long id;
     private Long productoId;
     private String nombreProducto;
-    private Double precioUnitario;
+    private BigDecimal precioUnitario;
     private Integer cantidad;
-    private Double subtotal;
+    private BigDecimal subtotal;
 }
